devops/requirements.txt
boto3==1.18.36
boto3-stubs==1.18.36
botocore==1.21.36
botocore-stubs==1.21.36
jmespath==0.10.0
mypy-boto3-ec2==1.18.36
mypy-boto3-iam==1.18.36
mypy-boto3-redshift==1.18.36
psycopg2-binary==2.9.1
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
